fixing master merge
Former-commit-id: ab84f1593caa813e2e75c69e927aeefd8df9b0ee [formerly 7f42b4be21087d02ce131a8e883efd560d7470bd] [formerly 82f60fa2b29cb4954dc67a6776f25d11cf9c5cdc [formerly c2e01f35684527beee3139c408dcb70d2f7b671b]]
Former-commit-id: 1ad37e33a15ae046e696e786466450e09ce6ee3d [formerly ce0995a43de50fc5d3057f4fd41f1cc31cd589ef]
Former-commit-id: 479d04a7ffcc6e181700e5631d15b17614f47a5f

+++ src/pop/individual.cc
@@ -426,7 +426,7 @@ namespace FT{
 
         #ifndef USE_CUDA
         // calculate program output matrix
-        MatrixXd Individual::out_trace(const Data& d, const Parameters& params, 
+        MatrixXf Individual::out_trace(const Data& d, const Parameters& params, 
                                        vector<Trace>& state_trace)
         {
             /*!
